Move 'struct vgmap' into its own include file
